add files for several rewards
