+++ pyproject.toml
@@ -52,6 +52,9 @@ data = [
 agent = [
     "agentscope"
 ]
+rm = [
+    "rm-gallery"
+]
 dev = [
     "pre-commit>=2.17.0",
     "black>=23.7.0",
