+++ trinity/common/rewards/__init__.py
@@ -1,11 +1,19 @@
 # -*- coding: utf-8 -*-
 """Reward functions for RFT"""
 
-from .basic_reward import REWARD_FUNCTIONS, AccuracyReward, FormatReward, RewardFn
+from .accuracy_reward import AccuracyReward
+from .countdown_reward import CountDownRewardFn
+from .format_reward import FormatReward
+from .math_reward import MathBoxedRewardFn, MathRewardFn
+from .reward_fn import REWARD_FUNCTIONS, RewardFn, RMGalleryFn
 
 __all__ = [
     "RewardFn",
+    "RMGalleryFn",
     "REWARD_FUNCTIONS",
     "AccuracyReward",
+    "CountDownRewardFn",
     "FormatReward",
+    "MathRewardFn",
+    "MathBoxedRewardFn",
 ]
